video integrado e diversas funcoes

+++ coopatos-care/backend/src/routes/index.ts
@@ -1,7 +1,14 @@
 import { Router } from "express";
 import { prisma } from "../prisma/client.js";
+import { v2 as cloudinary } from "cloudinary";
 
 const router = Router();
+
+cloudinary.config({
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET,
+});
 
 
 router.get("/", (req, res) => {
@@ -49,15 +56,15 @@ router.get("/reports", async (req, res) => {
 
 router.post("/reports", async (req, res) => {
   const {
-    employeeId,
-    categoryId,
-    description,
-    referencePoint,
-    latitude,
-    longitude,
-    address,
-    imageUrls,
-  } = req.body;
+  employeeId,
+  categoryId,
+  description,
+  referencePoint,
+  latitude,
+  longitude,
+  address,
+  mediaItems,
+} = req.body;
 
   console.log("BODY RECEBIDO:", req.body);
 
@@ -84,10 +91,12 @@ router.post("/reports", async (req, res) => {
       latitude,
       longitude,
 
-      images: imageUrls?.length
+      images: mediaItems?.length
   ? {
-      create: imageUrls.map((url: string) => ({
-        imageUrl: url,
+      create: mediaItems.map((item: any) => ({
+        imageUrl: item.imageUrl,
+        publicId: item.publicId,
+        resourceType: item.resourceType || "image",
       })),
     }
   : undefined,
@@ -141,6 +150,22 @@ router.patch("/reports/:id", async (req, res) => {
 router.delete("/report-images/:id", async (req, res) => {
   const id = Number(req.params.id);
 
+  const image = await prisma.reportImage.findUnique({
+    where: { id },
+  });
+
+  if (!image) {
+    return res.status(404).json({
+      error: "Imagem não encontrada.",
+    });
+  }
+
+  if (image.publicId) {
+    await cloudinary.uploader.destroy(image.publicId, {
+      resource_type: image.resourceType || "image",
+    });
+  }
+
   await prisma.reportImage.delete({
     where: { id },
   });
@@ -154,20 +179,22 @@ router.delete("/report-images/:id", async (req, res) => {
 
 router.post("/reports/:id/images", async (req, res) => {
   const reportId = Number(req.params.id);
-  const { imageUrls } = req.body;
+  const { mediaItems } = req.body;
 
-  if (!imageUrls || imageUrls.length === 0) {
+  if (!mediaItems || mediaItems.length === 0) {
     return res.status(400).json({
       error: "Nenhuma imagem enviada.",
     });
   }
 
   await prisma.reportImage.createMany({
-    data: imageUrls.map((url: string) => ({
-      reportId,
-      imageUrl: url,
-    })),
-  });
+  data: mediaItems.map((item: any) => ({
+    reportId,
+    imageUrl: item.imageUrl,
+    publicId: item.publicId,
+    resourceType: item.resourceType || "image",
+  })),
+});
 
   const report = await prisma.report.findUnique({
     where: { id: reportId },
